feat(services): add headertitle (#5)
### TL;DR

This update made changes in the `Services.jsx` and `services.astro` components, improving layout and style.

### What changed?

In `Services.jsx`, the CSS class was changed to adjust padding on the x-axis only. In the `services.astro` file, the layout and CSS classes were updated. Previously commented sections were uncommented and adjustments were made to the positioning and styling of elements. The changes focused on improving the look and feel of the 'Our Service' section.

### How to test?

To test the changes, run the application and navigate to the 'Services' page.  Observe the newly changed layout and style to ensure they appear correctly and as expected.

### Why make this change?

These changes were made to improve the user experience by providing a cleaner, more organized layout with improved visual design on the 'Services' page.

---